backend/requirements.txt
Flask==3.0.3
firebase-admin==6.6.0
gunicorn==21.2.0
python-dotenv==1.0.1
Flask-Cors==4.0.0
requests==2.32.3
pydub==0.25.1
cohere
grpcio-status==1.62.3
pytest
pytest-mock

